requirements.txt
pynput>=1.7.6
sounddevice>=0.4.6
soundfile>=0.12.1
openai>=1.0.0
faster-whisper>=0.10.0
pystray>=0.19.5
Pillow>=10.0.0
